fix(#55): allow trusted hosts on redirect
In order to avoid leaking sensitive information, the redirect plugin
does not forward sensitive headers unless the opts.Trusted field is
true. This commit adds an additional capability to also provide a list
of trusted host suffixes.

The opts.Trusted value will override any suffix list. Suffixes are not
perfect and care should be given to choose a value appropriate for the
given environment. For example, a 'trusted.com' suffix will match a host
that was 'untrusted.com.' A leading '.' can be used to avoid this, e.g.
'.trusted.com.'